bugfix: make repo's status change correct
